refactor: 차트 테이블 컴포넌트 명 변경
