+++ components/hospital/icu/main/chart/selected-chart/chart-body/table/chart-table-body/order-row-title.tsx
@@ -9,7 +9,7 @@ import type { IcuOrderColors, VitalRefRange } from '@/types/adimin'
 import type { SelectedIcuOrder } from '@/types/icu/chart'
 import { useCallback } from 'react'
 
-export default function CellsRowTitle({
+export default function OrderRowTitle({
   order,
   isSorting,
   index,
@@ -57,10 +57,9 @@ export default function CellsRowTitle({
     },
   })
 
-  const handleEditOrderDialogOpen = useCallback(
+  const handleClickOrderTitle = useCallback(
     (e: React.MouseEvent) => {
-      if (preview) return
-
+      // 오더 다중 선택시
       if (e.metaKey || e.ctrlKey) {
         e.preventDefault()
         setSelectedOrderPendingQueue((prev) => {
@@ -75,6 +74,8 @@ export default function CellsRowTitle({
         })
         return
       }
+
+      // 오더 수정하기 위해 다이얼로그 여는 경우
       reset()
       setOrderStep('upsert')
       setIsEditOrderMode(true)
@@ -122,7 +123,7 @@ export default function CellsRowTitle({
       <Button
         disabled={isOptimisticOrder}
         variant="ghost"
-        onClick={isSorting ? undefined : handleEditOrderDialogOpen}
+        onClick={isSorting ? undefined : handleClickOrderTitle}
         className={cn(
           'group flex h-11 justify-between rounded-none bg-transparent px-2 outline-none transition duration-300 hover:scale-[97%] hover:bg-transparent',
           isOptimisticOrder && 'animate-bounce',
